nvif: split out client interface definitions
